pyproject.toml
[project]
name = "gmail-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.160.0",
    "google-auth>=2.38.0",
    "google-auth-oauthlib>=1.2.1",
    "mcp[cli]>=1.2.1",
]
